bug figured out, it is simply broken, only use the catalog generated values
